dckr/similar-image-service/requirements.txt
flask==2.0.2
sentence-transformers~=2.2.0
ftfy~=6.1.1
python-dotenv~=0.19.2
requests~=2.27.1
Werkzeug~=2.0.3
Pillow~=9.0.1
flask-cors~=3.0.10